% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_dl.R
\name{bulk_dl}
\alias{bulk_dl}
\title{Bulk download full open FEMA data sets as .csv files}
\usage{
bulk_dl(data_set, output_dir = NULL, file_name = NULL, size_warning = TRUE)
}
\arguments{
\item{data_set}{A character string indicating the name of the data set
to download}

\item{output_dir}{An optional character string indicating the
directory (defaults to working directory)}

\item{file_name}{An optional character string indicating the file
name (defaults to data set name with time stamp)}

\item{size_warning}{A logical indicating whether to issue a warning
before proceeding with downloading a large file (default is TRUE)}
}
\value{
Returns a downloaded csv file of the data set to the
specified output directory.
}
\description{
Deprecated function for downloading full open FEMA data sets as .csv files.
}
\examples{
\dontrun{
bulk_dl("femaregions") # download the file
}
\dontrun{
file.remove("FemaRegions.csv") # clean up directory after file downloads
}
}
