% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_api_endpoints.R
\name{fema_api_endpoints}
\alias{fema_api_endpoints}
\title{Returns the API endpoint associated with a openFEMA data set}
\usage{
fema_api_endpoints(data_set)
}
\arguments{
\item{data_set}{A character string with the name of the data set to get the API endpoint for.}
}
\value{
Returns a character string containing the API endpoint url associated with the data set.
}
\description{
Returns the API endpoint associated with a openFEMA data set
}
\examples{
fema_api_endpoints("FimaNfipClaims")
fema_api_endpoints("fImAnfiPclaims")
fema_api_endpoints("fimanfippolicies")
}
