% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_data_fields.R
\name{fema_data_fields}
\alias{fema_data_fields}
\title{Get data fields and descriptions for a given FEMA data set}
\usage{
fema_data_fields(data_set)
}
\arguments{
\item{data_set}{a character string indicating the data set of interest}
}
\value{
Returns a tibble consisting of the data fields name,
along with information about each data field including the data type,
a description of the data field, and whether the data field is "searchable"
(i.e. can it be used to filter the returned data in an API query)
}
\description{
Get data fields and descriptions for a given FEMA data set
}
\examples{
\dontrun{
fema_data_fields("FimaNfipClaims")
}
\dontrun{
fema_data_fields("FimaNfipPolicies")
}
}
