% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_data_sets.R
\name{fema_data_sets}
\alias{fema_data_sets}
\title{Get a dataframe of avaliable FEMA data sets}
\usage{
fema_data_sets()
}
\value{
Returns a data frame containing information about each data set avaliable.
}
\description{
Get a dataframe of avaliable FEMA data sets
}
\examples{
fema_data_sets()
}
