% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_parameters.R
\name{valid_parameters}
\alias{valid_parameters}
\title{Get all valid API parameters for a given FEMA data set}
\usage{
valid_parameters(data_set = NULL)
}
\arguments{
\item{data_set}{A character string indicating the data set to return valid parameters for}
}
\value{
Returns a vector of parameter names that can be used to filter an API call for a given open FEMA dataset
}
\description{
Get all valid API parameters for a given FEMA data set
}
\examples{
valid_parameters("fimaNfipPolicies")
}
