% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_data_fields.R
\name{fema_data_fields}
\alias{fema_data_fields}
\title{Get data fields and descriptions for a given FEMA data set}
\usage{
fema_data_fields(data_set)
}
\arguments{
\item{data_set}{a character string indicating the data set of interest}
}
\value{
Returns a data frame consisting of the data field name, the data field title, the data type , a description of the data field, and whether it is searchable within an API query.
}
\description{
Get data fields and descriptions for a given FEMA data set
}
\examples{
fema_data_fields("FimaNfipClaims")
fema_data_fields("FimaNfipPolicies")
}
