% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_data_sets.R
\name{fema_data_sets}
\alias{fema_data_sets}
\title{Get a tibble of available FEMA data sets}
\usage{
fema_data_sets()
}
\value{
Returns a tibble containing meta data about each data
set available through the FEMA API. For more information see the FEMA
documentation page: https://www.fema.gov/about/openfema/data-sets.
}
\description{
Get a tibble of available FEMA data sets
}
\examples{
\dontrun{
fema_data_sets()
}
}
