% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema_data_sets.R
\name{fema_data_sets}
\alias{fema_data_sets}
\title{Get a dataframe of avaliable FEMA data sets}
\usage{
fema_data_sets(rfema_access = T)
}
\arguments{
\item{rfema_access}{logical indicating whether to return only the datasets that are accessable through the rfema package (default = TRUE).
Set to FALSE to get information on all data sets FEMA offers.}
}
\value{
Returns a data frame containing information about each data set avaliable.
}
\description{
Get a dataframe of avaliable FEMA data sets
}
\examples{
fema_data_sets(rfema_access  = TRUE)
fema_data_sets(rfema_access  = FALSE)
}
