% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_api_query.R
\name{gen_api_query}
\alias{gen_api_query}
\title{Helper function to generate the URL string for an API query}
\usage{
gen_api_query(data_set = NULL, top_n = 1000, filters = NULL, select = NULL)
}
\arguments{
\item{data_set}{character string specifiying the data set to query from}

\item{top_n}{numeric value indicating how many records to retun (default is 1000 which is the maximum number of records the API will return in a single query )}

\item{filters}{a list containing values of data fields that should be used to construct filters.}

\item{select}{a character vector of data fields to return (default is to return all data fields for the data set)}
}
\value{
The function returns a string containing a url that can be used to query the API.
}
\description{
Helper function to generate the URL string for an API query
}
\examples{
filter_list <- list(baseFloodElevation = c(5,6), countyCode = "34029" )
vars_to_select <- c("countyCode","baseFloodElevation")
url <- gen_api_query(data_set = "fimaNfipPolicies",top_n = 100, 
filters = filter_list, select = vars_to_select)
}
