% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_values.R
\name{parameter_values}
\alias{parameter_values}
\title{Returns examples of data field values}
\usage{
parameter_values(data_set = NULL, data_field = NULL)
}
\arguments{
\item{data_set}{a character vector specifying the data set}

\item{data_field}{a character vector specifying the data field}
}
\value{
returns a data frame with the unique data field values conatained in the first 1000 records of the full data set
}
\description{
Returns examples of data field values
}
\examples{
parameter_values(data_set = "fimaNfipPolicies",data_field = "crsClassCode")
}
